+++ bin/lib/log.js
@@ -1,0 +1,8 @@
+"use strict";
+
+// TODO: REMOVE? const log = (...args) => console.log(...args); // eslint-disable-line no-console
+const error = (...args) => console.error(...args); // eslint-disable-line no-console
+
+module.exports = {
+  error
+};
