+++ bin/trace-pkg.js
@@ -2,65 +2,10 @@
 
 "use strict";
 
-const fs = require("fs");
-const { promisify } = require("util");
+const { getArgs } = require("./lib/args");
+const { error } = require("./lib/log");
 
-const readFile = promisify(fs.readFile);
-
-const yargs = require("yargs/yargs");
-const yaml = require("js-yaml");
-
-const { version } = require("../package.json");
 const createPackage = require("../lib/actions/package").package;
-
-const NAME = "trace-pkg";
-
-// ============================================================================
-// Helpers
-// ============================================================================
-// TODO: REMOVE? const log = (...args) => console.log(...args); // eslint-disable-line no-console
-const error = (...args) => console.error(...args); // eslint-disable-line no-console
-
-const getArgs = async (args) => {
-  args = args || yargs.hideBin(process.argv);
-
-  // Parse
-  const parsed = yargs(args)
-    .usage(`Usage: ${NAME} [options]`)
-    // Files
-    .option("config", {
-      alias: "c",
-      describe: "Path to configuration file",
-      type: "string"
-    })
-    // Logistical
-    .exitProcess(false)
-    .help().alias("help", "h")
-    .version(version).alias("version", "v")
-    .strict();
-
-  // Validate
-  const { argv } = parsed;
-  const opts = {
-    help: !!argv.help,
-    version: !!argv.version
-  };
-
-  // Convert config file to full object.
-  if (typeof argv.config !== "undefined") {
-    try {
-      opts.config = yaml.safeLoad(await readFile(argv.config));
-    } catch (err) {
-      // Enhance message.
-      err.message = `Failed to load --config file"${argv.config}" with error: ${err.message}`;
-      throw err;
-    }
-  }
-
-  return {
-    opts
-  };
-};
 
 // ============================================================================
 // Script
@@ -81,6 +26,5 @@ if (require.main === module) {
 }
 
 module.exports = {
-  NAME,
   cli
 };
